requirements.txt
autogen==0.8.7
anthropic==0.42.0
fastapi==0.115.12
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-core==2.4.3
google-cloud-translate==3.20.2
googleapis-common-protos==1.70.0
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.54
langchain-openai==0.3.14
langchain-text-splitters==0.3.8
langsmith==0.3.32
lingua-language-detector==2.1.0
numpy==2.2.4
openai==1.75.0
pyautogen==0.8.7
pydantic==2.11.3
pydantic_core==2.33.1
pydantic-settings==2.9.1
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
youtube-transcript-api==1.0.3